Amélioration du HTML/CSS
Amélioration des tableaux, rajout d'un div.
Nettoyage  du CSS.
Nouveau menu gauche.
Nouvel index.
Sécurisation des pages faisait appel à $_REQUEST grace à if(isset()).

+++ supprimerLesInscritsAUneVisite.php
@@ -3,28 +3,39 @@
 include("_debut.inc.php");
 
 // ANNULER UNE VISITE
-
-$idVisite=$_REQUEST['idVisite'];
-// Cas 1ère étape (on vient de listeVisiteursAPrevenir.php)
-
-if ($_REQUEST['action']=='suppressionDemandee')
+if(isset($_REQUEST['idVisite'])) 
 {
-   echo "
-   <br><center><h5>Souhaitez-vous vraiment supprimer les inscrits ?
-   <br><br>
-   <a href='supprimerLesInscritsAUnevisite.php?action=suppressionConfirmee&idVisite=$idVisite'>
-   Oui</a>&nbsp; &nbsp; &nbsp; &nbsp;
-   <a href='listeVisiteursAPrevenir.php?idVisite=$idVisite'>Non</a></h5></center>";
+	$idVisite=$_REQUEST['idVisite'];
+	// Cas 1Ã¨re Ã©tape (on vient de listeVisiteursAPrevenir.php)
+
+	if ($_REQUEST['action']=='suppressionDemandee')
+	{
+	   echo "
+		<div id='corps'>
+			<h5>Souhaitez-vous vraiment supprimer les inscrits ?<br>
+			<a href='supprimerLesInscritsAUnevisite.php?action=suppressionConfirmee&idVisite=$idVisite'>
+			Oui</a>&nbsp; &nbsp; &nbsp; &nbsp; <a href='listeVisiteursAPrevenir.php?idVisite=$idVisite'>Non</a>
+			</h5>
+		</div>";
+	}
+
+	// Cas 2Ã¨me Ã©tape (on vient de demander la suppression des inscrits
+
+	if ($_REQUEST['action']=='suppressionConfirmee')
+	{
+	  $req = "delete from inscription where inscription.idVisite = $idVisite";
+	 mysql_query($req, $connexion);
+	   echo "
+		<div id='corps'>
+			<h5>Les visiteurs ont &eacute;t&eacute; supprim&eacute;s</h5>
+			<a href='listeVisitesPourDetail.php'>Retour</a>
+		</div>";
+	}
 }
-
-// Cas 2ème étape (on vient de demander la suppression des inscrits
-
-if ($_REQUEST['action']=='suppressionConfirmee')
-{
-  $req = "delete from inscription where inscription.idVisite = $idVisite";
- mysql_query($req, $connexion);
-   echo "
-   <br><br><center><h5>Les visiteurs ont &eacute;t&eacute; supprim&eacute;s</h5>
-
-   <a href='listeVisitesPourDetail.php'>Retour</a></center>";
+else {
+	echo "<div id='corps'><p>Aucune valeur dans la superglobale.</p></div>";
 }
+?>
+
+</body>
+</html>
